pyproject.toml
[project]
name = "data-augmentation-with-llms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.4.1",
    "google-genai>=1.52.0",
    "hydra-core>=1.3.2",
    "openai>=1.96.1",
    "tqdm>=4.67.1",
    "transformers>=4.57.1",
    "xai-sdk>=1.4.0",
]
